limit charge now range to minAmpsPerTWC
Lowest value for charge now should be minAmpsPerTWC, as below this value
we do not start charging. This would confuse a user trying to charge
with 5 amps, or any other value below minAmpsPerTWC.
While we are at it, set the highest value to minAmpsPerTWC, in case
wiringMaxAmpsPerTWC is not set for some reason. This avoids an empty
ampsList. We might as well want to throw an Exception in such case,
though.